+++ src/event.rs
@@ -41,8 +41,8 @@ pub enum FromEventError {
     /// couldn't parse request from interface
     #[error("couldn't parse request from interface {0}")]
     Interface(String),
-    /// object has wrong base path
-    #[error("object {interface} has wrong base path {base_path}")]
+    /// couldn't parse the event path
+    #[error("the interface {interface} has wrong path {base_path}")]
     Path {
         /// Interface that generated the error
         interface: &'static str,
@@ -60,6 +60,14 @@ pub enum FromEventError {
     /// object data passed to individual
     #[error("object data passed to individual {interface}{endpoint}")]
     Object {
+        /// Interface that generated the error
+        interface: &'static str,
+        /// endpoint
+        endpoint: String,
+    },
+    /// unset passed to endpoint without allow unset
+    #[error("unset passed to {interface}{endpoint} without allow unset")]
+    Unset {
         /// Interface that generated the error
         interface: &'static str,
         /// endpoint
@@ -207,6 +215,8 @@ mod tests {
             Luminosity(i32),
             #[mapping(endpoint = "/sensor/temperature")]
             Temperature(f64),
+            #[mapping(endpoint = "/sensor/unsettable", allow_unset = true)]
+            Unsettable(Option<bool>),
         }
 
         let event = DeviceEvent {
@@ -230,6 +240,30 @@ mod tests {
         let temperature = Sensor::from_event(event).expect("couldn't parse the event");
 
         let expected = Sensor::Temperature(3.);
+
+        assert_eq!(temperature, expected);
+
+        let event = DeviceEvent {
+            interface: "com.example.Sensor".to_string(),
+            path: "/sensor/unsettable".to_string(),
+            data: Value::Individual(AstarteType::Boolean(true)),
+        };
+
+        let temperature = Sensor::from_event(event).expect("couldn't parse the event");
+
+        let expected = Sensor::Unsettable(Some(true));
+
+        assert_eq!(temperature, expected);
+
+        let event = DeviceEvent {
+            interface: "com.example.Sensor".to_string(),
+            path: "/sensor/unsettable".to_string(),
+            data: Value::Unset,
+        };
+
+        let temperature = Sensor::from_event(event).expect("couldn't parse the event");
+
+        let expected = Sensor::Unsettable(None);
 
         assert_eq!(temperature, expected);
     }