+++ astarte-device-sdk-derive/CHANGELOG.md
@@ -11,6 +11,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Implement the FromEvent derive macro for individual interfaces, via the `aggregation` attribute to
   the macro [#375](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/375)
+- Add the `allow_unset` attribute to permit `Option` values for `Value::Unset`
+  [#378](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/378)
 
 ## [0.8.4] - 2024-09-11